+++ spec/support/normalize_indent.rb
@@ -3,7 +3,7 @@ module IndentNormalizer
   # for writing code without having the leading indentation count.
   def normalize_indent(code)
     leading_indent = code[/^(\s*)/, 1]
-    code.strip.gsub(/\n#{leading_indent}/, "\n")
+    code.lstrip.gsub(/\n#{leading_indent}/, "\n")
   end
 end